// task-dispatch/src/app/models/taskclass.ts
export interface ITaskClass {
  functionalAreaId: number;
  sortByStatTask: Boolean;
  classBrief: String;
  isStatDisplayed: Boolean;
  classID: number;
  forceField: Boolean;
  required: Boolean;
  header: String;
  customHeader: String;
  textLength: Number;
  controlName: string;
  controlType: String;
  pickListSource: String;
  isSiteSpecific: Boolean;
  displayOrder: Number;
  enteredBy: Number;
  enteredAt: String;
  errorCode: Number;
  errorResponse: String;
}